requirements.txt
scipy==1.2.0
numpy==1.18.4
opencv_python==4.2.0.34
lmdb==0.98
seaborn==0.9.0
six==1.15.0
tensorflow==1.15.2
torch==1.4.0
RoboNet.egg==info
scikit_image==0.16.2
easydict==1.9
absl_py==0.8.0
gdown==3.12.2
baselines.egg==info
tqdm==4.46.1
imageio==2.9.0
matplotlib==3.1.1
pandas==0.25.2
absl==0.0
baselines==0.1.5
ipython==7.19.0
Pillow==8.0.1
scikit_learn==0.23.2
scikit-image
torchvision==0.8.1
